Apply normalisation on a data frame using specific Lambda value

Apply normalisation on a data frame using specific Lambda value

Usage,

normaliseDataInfo(dat, lambdas)

Arguments

dat

a data frame containing the variables.

lambdas

a numeric vector from the boxcox function

Value

a list of three elements, the first element, lambdas, is a numeric vector containing lambdas, the second element, scaled, is a numeric vector containing scale adjustment, the last element, transDat, is a data frame comprising the transformed variables.

References

Courtney, M. G. R., and Chang, K. C. (2018) Dealing with non-normality: an introduction and step-by-step guide using R. TEST, doi: 10.1111/test.12154.

Examples

if (FALSE) normaliseDataInfo(mtcars, getLambda(mtcars, parallel = FALSE))
normaliseDataInfo(mtcars, getLambda(mtcars, parallel = FALSE))
#> $lambdas
#>   mpg   cyl  disp    hp  drat    wt  qsec    vs    am  gear  carb 
#>  0.60  1.23  0.55  0.63  0.28  0.65  0.03 -1.09 -1.66 -0.99  0.27 
#> 
#> $scaled
#>         mpg         cyl        disp          hp        drat          wt 
#>   5.5874699   0.5402412  20.9657459  10.5119843   9.2595409   2.3108887 
#>        qsec          vs          am        gear        carb 
#> 561.6057344   1.8859433   1.4629275  14.9509632   9.2935391 
#> 
#> $transDat
#>                          mpg      cyl      disp        hp     drat       wt
#> Mazda RX4           22.34358 5.922063 194.10774 128.42726 4.010692 2.810257
#> Mazda RX4 Wag       22.34358 5.922063 194.10774 128.42726 4.010692 3.079211
#> Datsun 710          24.09960 4.000000 127.65966 108.04090 3.961808 2.481787
#> Hornet 4 Drive      22.73885 5.922063 296.83448 128.42726 3.143788 3.425168
#> Hornet Sportabout   20.00945 8.000000 386.22501 197.44680 3.223876 3.647076
#> Valiant             19.38196 5.922063 264.60189 122.56056 2.760000 3.666552
#> Duster 360          15.19495 8.000000 386.22501 261.70545 3.291509 3.772971
#> Merc 240D           25.61445 4.000000 178.17627  66.84686 3.802242 3.400182
#> Merc 230            24.09960 4.000000 170.90146 110.50695 4.030119 3.360059
#> Merc 280            20.52622 5.922063 202.94404 143.23683 4.030119 3.647076
#> Merc 280C           19.06511 5.922063 202.94404 143.23683 4.030119 3.647076
#> Merc 450SE          17.55724 8.000000 313.44903 202.31983 3.132240 4.242959
#> Merc 450SL          18.53222 8.000000 313.44903 202.31983 3.132240 3.925737
#> Merc 450SLC         16.22327 8.000000 313.44903 202.31983 3.132240 3.973003
#> Cadillac Fleetwood  10.40000 8.000000 472.00000 225.97254 2.967653 5.278556
#> Lincoln Continental 10.40000 8.000000 463.28458 235.13105 3.050638 5.424000
#> Chrysler Imperial   15.65509 8.000000 448.52817 248.57869 3.313851 5.358171
#> Fiat 128            32.66039 4.000000  83.66873  72.53322 4.183024 2.346280
#> Honda Civic         30.97125 4.000000  78.77563  52.00000 4.930000 1.644023
#> Toyota Corolla      33.90000 4.000000  71.10000  71.12388 4.313312 1.917158
#> Toyota Corona       22.83721 4.000000 144.22111 112.95387 3.812359 2.642294
#> Dodge Challenger    16.56060 8.000000 351.01113 172.26035 2.760000 3.724743
#> AMC Javelin         16.22327 8.000000 338.81329 172.26035 3.223876 3.642201
#> Camaro Z28          14.02149 8.000000 378.01621 261.70545 3.842593 4.029434
#> Pontiac Firebird    20.52622 8.000000 418.07882 197.44680 3.143788 4.034123
#> Fiat X1-9           28.26239 4.000000  84.15339  72.53322 4.183024 2.037484
#> Porsche 914-2       27.09004 4.000000 144.48838 105.55515 4.503031 2.277559
#> Lotus Europa        30.97125 4.000000 109.05389 131.89991 3.882634 1.513000
#> Ford Pantera L      16.89532 8.000000 378.84180 277.91331 4.313312 3.380143
#> Ferrari Dino        21.03764 5.922063 176.09386 197.44680 3.730862 2.969520
#> Maserati Bora       15.99691 8.000000 336.16681 335.00000 3.648058 3.772971
#> Volvo 142E          22.73885 4.000000 145.42225 127.26195 4.211211 2.980029
#>                         qsec vs am    gear     carb
#> Mazda RX4           16.81892  0  1 4.24879 5.219011
#> Mazda RX4 Wag       17.43231  0  1 4.24879 5.219011
#> Datsun 710          19.07276  1  1 4.24879 1.000000
#> Hornet 4 Drive      19.87583  1  0 3.00000 1.000000
#> Hornet Sportabout   17.43231  0  0 3.00000 2.912683
#> Valiant             20.60076  1  0 3.00000 1.000000
#> Duster 360          16.11574  0  0 3.00000 5.219011
#> Merc 240D           20.39908  1  0 4.24879 2.912683
#> Merc 230            22.90000  1  0 4.24879 2.912683
#> Merc 280            18.76388  1  0 4.24879 5.219011
#> Merc 280C           19.35723  1  0 4.24879 5.219011
#> Merc 450SE          17.83749  0  0 3.00000 4.209163
#> Merc 450SL          18.04731  0  0 3.00000 4.209163
#> Merc 450SLC         18.46009  0  0 3.00000 4.209163
#> Cadillac Fleetwood  18.43966  0  0 3.00000 5.219011
#> Lincoln Continental 18.27545  0  0 3.00000 5.219011
#> Chrysler Imperial   17.85858  0  0 3.00000 5.219011
#> Fiat 128            19.90423  1  1 4.24879 1.000000
#> Honda Civic         18.98361  1  1 4.24879 2.912683
#> Toyota Corolla      20.30669  1  1 4.24879 1.000000
#> Toyota Corona       20.40830  1  0 3.00000 1.000000
#> Dodge Challenger    17.26995  0  0 3.00000 2.912683
#> AMC Javelin         17.73170  0  0 3.00000 2.912683
#> Camaro Z28          15.61219  0  0 3.00000 5.219011
#> Pontiac Firebird    17.46462  0  0 3.00000 2.912683
#> Fiat X1-9           19.35723  1  1 4.24879 1.000000
#> Porsche 914-2       17.08425  0  1 5.00000 2.912683
#> Lotus Europa        17.30254  1  1 5.00000 2.912683
#> Ford Pantera L      14.50000  0  1 5.00000 5.219011
#> Ferrari Dino        15.71870  0  1 5.00000 6.782316
#> Maserati Bora       14.62548  0  1 5.00000 8.000000
#> Volvo 142E          19.06288  1  1 4.24879 2.912683
#>